move FilesContext to top level of Files directory
